refactor(cli): move ACP contracts out of serve
