+++ packages/cli/src/runtime/workspace-skills-mapping.test.ts
