+++ packages/cli/src/serve/routes/session.ts
@@ -34,8 +34,8 @@ import type { SessionArtifactInput } from '@qwen-code/acp-bridge/sessionArtifact
 import { parseSessionSource } from '@qwen-code/acp-bridge';
 import type { Application, Request, RequestHandler, Response } from 'express';
 import { writeStderrLine } from '../../utils/stdioHelpers.js';
-import { isChannelDeliveryError } from '../channel-delivery-ipc.js';
-import { parseChannelDelivery } from '../channel-delivery.js';
+import { isChannelDeliveryError } from '../../runtime/channel-delivery-ipc.js';
+import { parseChannelDelivery } from '../../runtime/channel-delivery.js';
 import {
   canonicalizeWorkspace,
   InvalidClientIdError,
